# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Yamo Laundry API
  version: 1.0.1
  description: >-
    REST API for the Yamo laundry platform. Includes customer, orders,
    tasks, billing, geo, notifications, livreur management and live tracking.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api.yamo.example
    description: Production
...
paths:
...
  /api/livreurs/{id}/location:
    post:
      tags: [Tracking]
      summary: Post livreur location update
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LocationUpdate' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/LivreurLocation' } } } }
  /api/livreurs/{id}/location/recent:
    get:
      tags: [Tracking]
      summary: Get recent livreur locations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: minutes
          required: false
          schema: { type: integer, format: int64, default: 30 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LivreurLocation' }
  /api/orders/{orderId}/livreur/latest:
    get:
      tags: [Tracking]
      summary: Get latest location for order's livreur
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/LivreurLocation' } } } }
        '204': { description: No content }
  /api/orders/{orderId}/livreur/stream:
    get:
      tags: [Tracking]
      summary: Subscribe to SSE location stream for order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
...
    LivreurRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
    LivreurLocation:
      type: object
      properties:
        id: { type: integer, format: int64 }
        livreur: { $ref: '#/components/schemas/LivreurRef' }
        order: { $ref: '#/components/schemas/OrderRef' }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        heading: { type: number, format: double }
        speed: { type: number, format: double }
        capturedAt: { type: string, format: date-time }
    LocationUpdate:
      type: object
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        heading: { type: number, format: double }
        speed: { type: number, format: double }
        orderId: { type: integer, format: int64 }
...
security:
  - bearerAuth: []